Build Acts Component
The acts component now displays a table with edit, save, and delete
functionality for each row. It hooks into the ActService to query the API
with the actions on the user interface. Each model should implement a
component in the same fashion as what has been done here for Acts.