Provide Creation Form for Entities
Under the table view of each entity is a button that opens an in-line form to allow the user to add a new entity to the database. This has not been implemented for the Shows or Venue entities as a decision on datetime formatting must be made.

Included in this commit is also a modification to entity model files. Each was given a constructor to assist in proper data binding.